Update home page layout and styles
The home page has been improved to show a list of the most active boards and categories. The layout of the home page has been updated to make better use of the space available, and CSS styles have been added to improve presentation. Additionally, a minor change has been made in the CategoryRepository to ensure the ID of each category is included in the query results.